Fabric: do `set_epoch` for `batch_sampler.sampler` (#16841)
